1. receive private msg using signal r - done 2. search in pay - done

+++ Sourcecode/Mobile/Strive Greeter/Greeter/DTOs/Model.cs
@@ -1149,7 +1149,7 @@ namespace Greeter.DTOs
         public string SenderLastName { get; set; }
 
         [JsonProperty("ReceipientId")]
-        public int ReceipientID { get; set; }
+        public long ReceipientID { get; set; }
 
         [JsonProperty("RecipientFirstName")]
         public string RecipientFirstName { get; set; }
